requirements.txt
langchain
langchain-groq
pydantic
pandas
python-dotenv
streamlit
